+++ packages/amplify-cli-core/src/feature-flags/featureFlags.ts
@@ -644,13 +644,13 @@ export class FeatureFlags {
         name: 'enableAutoIndexQueryNames',
         type: 'boolean',
         defaultValueForExistingProjects: false,
-        defaultValueForNewProjects: false,
+        defaultValueForNewProjects: true,
       },
       {
         name: 'respectPrimaryKeyAttributesOnConnectionField',
         type: 'boolean',
         defaultValueForExistingProjects: false,
-        defaultValueForNewProjects: false,
+        defaultValueForNewProjects: true,
       },
       {
         name: 'shouldDeepMergeDirectiveConfigDefaults',