+++ packages/amplify-cli-core/src/overrides-manager/override-skeleton-generator.ts
@@ -1,11 +1,15 @@
 import { printer, prompter } from 'amplify-prompts';
 import execa from 'execa';
 import * as fs from 'fs-extra';
-import { EOL } from 'os';
 import * as path from 'path';
-import { $TSAny, $TSContext, getPackageManager, pathManager } from '../index';
+import {
+  $TSAny, $TSContext, AmplifyError, getPackageManager, pathManager,
+} from '../index';
 import { JSONUtilities } from '../jsonUtilities';
 
+/**
+ * This method generates the default/template overrides file
+ */
 export const generateOverrideSkeleton = async (context: $TSContext, srcResourceDirPath: string, destDirPath: string): Promise<void> => {
   // 1. Create skeleton package
   const backendDir = pathManager.getBackendDirPath();
@@ -32,7 +36,13 @@ export const generateOverrideSkeleton = async (context: $TSContext, srcResourceD
   }
 };
 
-export async function buildOverrideDir(cwd: string, destDirPath: string): Promise<boolean> {
+/**
+ * Returns true if a Typescript overrides file is found, and compiled successfully into an overrides.js file.
+ * Returns false if no Typescript overrides file is found.
+ *
+ * Throws if a Typescript overrides file is found, but does not compile.
+ */
+export const buildOverrideDir = async (cwd: string, destDirPath: string): Promise<boolean> => {
   const overrideFileName = path.join(destDirPath, 'override.ts');
   if (!fs.existsSync(overrideFileName)) {
     // return when no override file found
@@ -52,7 +62,10 @@ export async function buildOverrideDir(cwd: string, destDirPath: string): Promis
   const packageManager = getPackageManager(cwd);
 
   if (packageManager === null) {
-    throw new Error('No package manager found. Please install npm or yarn to compile overrides for this project.');
+    throw new AmplifyError('MissingOverridesInstallationRequirementsError', {
+      message: 'No package manager found.',
+      resolution: 'Please install npm or yarn to compile overrides for this project.',
+    });
   }
 
   try {
@@ -76,7 +89,10 @@ export async function buildOverrideDir(cwd: string, destDirPath: string): Promis
     const localTscExecutablePath = path.join(cwd, 'node_modules', '.bin', 'tsc');
 
     if (!fs.existsSync(localTscExecutablePath)) {
-      throw new Error('Typescript executable not found. Please add it as a dev-dependency in the package.json file for this resource.');
+      throw new AmplifyError('MissingOverridesInstallationRequirementsError', {
+        message: 'Typescript executable not found.',
+        resolution: 'Please add it as a dev-dependency in the package.json file for this resource.',
+      });
     }
     execa.sync(localTscExecutablePath, [`--project`, `${tsConfigDestFilePath}`], {
       cwd: tsConfigDir,
@@ -86,17 +102,26 @@ export async function buildOverrideDir(cwd: string, destDirPath: string): Promis
     return true;
   } catch (error: $TSAny) {
     if (error.code === 'ENOENT') {
-      throw new Error(`Packaging overrides failed. Could not find ${packageManager} executable in the PATH.`);
+      throw new AmplifyError('MissingOverridesInstallationRequirementsError', {
+        message: `Packaging overrides failed. Could not find ${packageManager} executable in the PATH.`,
+      });
     } else {
-      throw new Error(`Packaging overrides failed with the error:${EOL}${error.message}`);
+      throw new AmplifyError('InvalidOverrideError', {
+        message: `Packaging overrides failed.`,
+        details: error.message,
+        resolution: 'There may be errors in your overrides file. If so, fix the errors and try again.',
+      }, error);
     }
   }
-}
+};
 
-export const generateAmplifyOverrideProjectBuildFiles = (backendDir: string, srcResourceDirPath: string) => {
+/**
+ * this method adds the package.json & tsconfig.json files needed for overrides
+ */
+export const generateAmplifyOverrideProjectBuildFiles = (backendDir: string, srcResourceDirPath: string): void => {
   const packageJSONFilePath = path.join(backendDir, 'package.json');
   const tsConfigFilePath = path.join(backendDir, 'tsconfig.json');
-  // add package.json to amplofy backend
+  // add package.json to amplify backend
   if (!fs.existsSync(packageJSONFilePath)) {
     const packageJson = JSONUtilities.readJson(path.join(srcResourceDirPath, 'package.json'));
     JSONUtilities.writeJson(packageJSONFilePath, packageJson);
@@ -109,7 +134,10 @@ export const generateAmplifyOverrideProjectBuildFiles = (backendDir: string, src
   }
 };
 
-export const generateTsConfigforProject = (srcResourceDirPath: string, destDirPath: string) => {
+/**
+ * this method generates the tsconfig file template for overrides
+ */
+export const generateTsConfigforProject = (srcResourceDirPath: string, destDirPath: string): void => {
   const overrideFileName = path.join(destDirPath, 'override.ts');
   // ensure build dir path
   fs.ensureDirSync(path.join(destDirPath, 'build'));