+++ packages/amplify-e2e-core/src/index.ts
@@ -1,3 +1,8 @@
+/* eslint-disable prefer-arrow/prefer-arrow-functions */
+/* eslint-disable func-style */
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
+/* eslint-disable jsdoc/require-jsdoc */
+/* eslint-disable import/no-cycle */
 import * as os from 'os';
 import * as path from 'path';
 import * as fs from 'fs-extra';
@@ -12,12 +17,12 @@ export * from './configure';
 export * from './init';
 export * from './utils';
 export * from './categories';
-export * from './utils/sdk-calls';
 export * from './export';
 export { addFeatureFlag } from './utils/feature-flags';
 export * from './cli-version-controller';
 
 declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace NodeJS {
     interface Global {
       getRandomId: () => string;
@@ -27,9 +32,6 @@ declare global {
 
 const amplifyTestsDir = 'amplify-e2e-tests';
 
-/**
- *
- */
 export function getCLIPath(testingWithLatestCodebase = false) {
   if (!testingWithLatestCodebase) {
     if (process.env.AMPLIFY_PATH && fs.existsSync(process.env.AMPLIFY_PATH)) {
@@ -43,16 +45,10 @@ export function getCLIPath(testingWithLatestCodebase = false) {
   return amplifyScriptPath;
 }
 
-/**
- *
- */
 export function isTestingWithLatestCodebase(scriptRunnerPath) {
   return scriptRunnerPath === process.execPath;
 }
 
-/**
- *
- */
 export function getScriptRunnerPath(testingWithLatestCodebase = false) {
   if (!testingWithLatestCodebase) {
     return process.platform === 'win32' ? 'node.exe' : 'exec';
@@ -62,9 +58,6 @@ export function getScriptRunnerPath(testingWithLatestCodebase = false) {
   return process.execPath;
 }
 
-/**
- *
- */
 export function getNpxPath() {
   let npxPath = 'npx';
   if (process.platform === 'win32') {
@@ -73,9 +66,6 @@ export function getNpxPath() {
   return npxPath;
 }
 
-/**
- *
- */
 export function getNpmPath() {
   let npmPath = 'npm';
   if (process.platform === 'win32') {
@@ -84,16 +74,6 @@ export function getNpmPath() {
   return npmPath;
 }
 
-/**
- *
- */
-export function isCI(): boolean {
-  return !!(process.env.CI && process.env.CIRCLECI);
-}
-
-/**
- *
- */
 export function injectSessionToken(profileName: string) {
   const credentialsContents = ini.parse(fs.readFileSync(pathManager.getAWSCredentialsFilePath()).toString());
   credentialsContents[profileName] = credentialsContents[profileName] || {};
@@ -101,16 +81,10 @@ export function injectSessionToken(profileName: string) {
   fs.writeFileSync(pathManager.getAWSCredentialsFilePath(), ini.stringify(credentialsContents));
 }
 
-/**
- *
- */
 export function npmInstall(cwd: string) {
   spawnSync('npm', ['install'], { cwd });
 }
 
-/**
- *
- */
 export async function installAmplifyCLI(version = 'latest') {
   spawnSync('npm', ['install', '-g', `@aws-amplify/cli@${version}`], {
     cwd: process.cwd(),
@@ -122,11 +96,9 @@ export async function installAmplifyCLI(version = 'latest') {
     : path.join(os.homedir(), '.npm-global', 'bin', 'amplify');
 }
 
-/**
- *
- */
 export async function createNewProjectDir(
   projectName: string,
+  // eslint-disable-next-line spellcheck/spell-checker
   prefix = path.join(fs.realpathSync(os.tmpdir()), amplifyTestsDir),
 ): Promise<string> {
   const currentHash = execSync('git rev-parse --short HEAD', { cwd: __dirname }).toString().trim();
@@ -141,10 +113,8 @@ export async function createNewProjectDir(
   return projectDir;
 }
 
-/**
- *
- */
 export const createTempDir = () => {
+  // eslint-disable-next-line spellcheck/spell-checker
   const osTempDir = fs.realpathSync(os.tmpdir());
   const tempProjectDir = path.join(osTempDir, amplifyTestsDir, uuid());