+++ packages/amplify-provider-awscloudformation/src/root-stack-builder/root-stack-transform.ts
@@ -1,7 +1,7 @@
 import { AmplifyRootStackTemplate } from '@aws-amplify/cli-extensibility-helper';
 import * as cdk from '@aws-cdk/core';
 import {
-  $TSContext, AmplifyFault, buildOverrideDir, CFNTemplateFormat, pathManager, Template, writeCFNTemplate,
+  $TSContext, AmplifyError, AmplifyFault, buildOverrideDir, CFNTemplateFormat, pathManager, Template, writeCFNTemplate,
 } from 'amplify-cli-core';
 import { formatter } from 'amplify-prompts';
 import * as fs from 'fs-extra';
@@ -71,8 +71,15 @@ export class AmplifyRootStackTransform {
           external: true,
         },
       });
-
-      sandboxNode.run(overrideCode).override(this._rootTemplateObj as AmplifyRootStackTemplate);
+      try {
+        await sandboxNode.run(overrideCode).override(this._rootTemplateObj as AmplifyRootStackTemplate);
+      } catch (err) {
+        throw new AmplifyError('InvalidOverrideError', {
+          message: `Executing overrides failed.`,
+          details: err.message,
+          resolution: 'There may be runtime errors in your overrides file. If so, fix the errors and try again.',
+        }, err);
+      }
     }
   };