+++ packages/amplify-category-hosting/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@3.4.4...amplify-category-hosting@3.4.5) (2022-11-17)
+
+**Note:** Version bump only for package amplify-category-hosting
+
+
+
+
+
 ## [3.4.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@3.4.3...amplify-category-hosting@3.4.4) (2022-10-27)
 
 **Note:** Version bump only for package amplify-category-hosting